+++ src/lib/pages/account-details/index.tsx
@@ -1,4 +1,5 @@
 import {
+  Box,
   Flex,
   Heading,
   TabList,
@@ -8,15 +9,32 @@ import {
   Text,
 } from "@chakra-ui/react";
 import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
 
 import { BackButton } from "lib/components/button";
 import { CustomTab } from "lib/components/CustomTab";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import PageContainer from "lib/components/PageContainer";
 import { InvalidState } from "lib/components/state/InvalidState";
+import { ViewMore } from "lib/components/table/ViewMore";
 import { useValidateAddress } from "lib/hooks";
+import { useAccountDetailsTableCounts } from "lib/model/account";
+import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 import type { HumanAddr } from "lib/types";
 import { getFirstQueryParam } from "lib/utils";
+
+import { ContractTable } from "./components/tables/contracts";
+
+enum TabIndex {
+  Overview,
+  Delegations,
+  Assets,
+  Txs,
+  Codes,
+  Contracts,
+  Admins,
+  Proposals,
+}
 
 interface AccountDetailsBodyProps {
   accountAddress: HumanAddr;
@@ -25,7 +43,16 @@ interface AccountDetailsBodyProps {
 const InvalidAccount = () => <InvalidState title="Account does not exist" />;
 
 const AccountDetailsBody = ({ accountAddress }: AccountDetailsBodyProps) => {
+  const [tabIndex, setTabIndex] = useState(0);
   const tableHeaderId = "accountDetailsTab";
+  const {
+    tableCounts,
+    // refetchCodes,
+    // refetchContractsAdminCount,
+    refetchContractsCount,
+    // refetchCountTxs,
+    // refetchProposalCount,
+  } = useAccountDetailsTableCounts(accountAddress);
 
   return (
     <>
@@ -47,27 +74,58 @@ const AccountDetailsBody = ({ accountAddress }: AccountDetailsBodyProps) => {
           />
         </Flex>
       </Flex>
-      <Tabs>
+      <Tabs index={tabIndex}>
         <TabList
           borderBottom="1px solid"
           borderColor="pebble.700"
           id={tableHeaderId}
         >
-          <CustomTab>Overall</CustomTab>
+          <CustomTab onClick={() => setTabIndex(TabIndex.Overview)}>
+            Overall
+          </CustomTab>
           {/* TODO: add counts for Delegations */}
-          <CustomTab>Delegations</CustomTab>
+          <CustomTab onClick={() => setTabIndex(TabIndex.Delegations)}>
+            Delegations
+          </CustomTab>
           {/* TODO: add counts for Assets */}
-          <CustomTab count={0}>Assets</CustomTab>
-          {/* TODO: add counts for Txs */}
-          <CustomTab count={0}>Transactions</CustomTab>
-          {/* TODO: add counts for Codes */}
-          <CustomTab count={0}>Codes</CustomTab>
-          {/* TODO: add counts for Contracts */}
-          <CustomTab count={0}>Contracts</CustomTab>
-          {/* TODO: add counts for Admins */}
-          <CustomTab count={0}>Admins</CustomTab>
-          {/* TODO: add counts for Proposals */}
-          <CustomTab count={0}>Proposals</CustomTab>
+          <CustomTab count={0} onClick={() => setTabIndex(TabIndex.Assets)}>
+            Assets
+          </CustomTab>
+          <CustomTab
+            count={tableCounts.countTxs}
+            isDisabled={!tableCounts.countTxs}
+            onClick={() => setTabIndex(TabIndex.Txs)}
+          >
+            Transactions
+          </CustomTab>
+          <CustomTab
+            count={tableCounts.codesCount}
+            isDisabled={!tableCounts.codesCount}
+            onClick={() => setTabIndex(TabIndex.Codes)}
+          >
+            Codes
+          </CustomTab>
+          <CustomTab
+            count={tableCounts.contractsCount}
+            isDisabled={!tableCounts.contractsCount}
+            onClick={() => setTabIndex(TabIndex.Contracts)}
+          >
+            Contracts
+          </CustomTab>
+          <CustomTab
+            count={tableCounts.contractsAdminCount}
+            isDisabled={!tableCounts.contractsAdminCount}
+            onClick={() => setTabIndex(TabIndex.Admins)}
+          >
+            Admins
+          </CustomTab>
+          <CustomTab
+            count={tableCounts.proposalCount}
+            isDisabled={!tableCounts.proposalCount}
+            onClick={() => setTabIndex(TabIndex.Proposals)}
+          >
+            Proposals
+          </CustomTab>
         </TabList>
         <TabPanels>
           <TabPanel p={0}>
@@ -79,8 +137,19 @@ const AccountDetailsBody = ({ accountAddress }: AccountDetailsBodyProps) => {
             <Text>Transactions</Text>
             {/* TODO: replace with the truncated Codes table */}
             <Text>Stored Codes</Text>
-            {/* TODO: replace with the truncated Contracts table */}
-            <Text>Contract Instances</Text>
+            <Box>
+              <ContractTable
+                walletAddress={accountAddress}
+                scrollComponentId={tableHeaderId}
+                totalData={tableCounts.contractsCount}
+                refetchCount={refetchContractsCount}
+                isPreview
+              />
+              {tableCounts.contractsCount &&
+                tableCounts.contractsCount >= 5 && (
+                  <ViewMore onClick={() => setTabIndex(TabIndex.Contracts)} />
+                )}
+            </Box>
             {/* TODO: replace with the truncated Admins table */}
             <Text>Contract Admin</Text>
             {/* TODO: replace with the truncated Proposals table */}
@@ -103,8 +172,12 @@ const AccountDetailsBody = ({ accountAddress }: AccountDetailsBodyProps) => {
             <Text>Stored Codes</Text>
           </TabPanel>
           <TabPanel p={0}>
-            {/* TODO: replace with the full Contracts table */}
-            <Text>Contract Instances</Text>
+            <ContractTable
+              walletAddress={accountAddress}
+              scrollComponentId={tableHeaderId}
+              totalData={tableCounts.contractsCount}
+              refetchCount={refetchContractsCount}
+            />
           </TabPanel>
           <TabPanel p={0}>
             {/* TODO: replace with the full Admins table */}
@@ -126,6 +199,10 @@ const AccountDetails = () => {
   const accountAddressParam = getFirstQueryParam(
     router.query.accountAddress
   ) as HumanAddr;
+
+  useEffect(() => {
+    if (router.isReady) AmpTrack(AmpEvent.TO_ACCOUNT_DETAIL);
+  }, [router.isReady]);
 
   return (
     <PageContainer>