+++ src/lib/services/amplitude.tsx
@@ -40,6 +40,7 @@ export enum AmpEvent {
   TO_LIST_OTHERS = "To List Others",
   TO_ALL_LISTS = "To All Lists",
   TO_ALL_PROJECTS = "To All Public Projects",
+  TO_ACCOUNT_DETAIL = "To Account Detail",
   TO_CONTRACT_DETAIL = "To Contract Detail",
   TO_CODE_DETAIL = "To Code Detail",
   TO_PROJECT_DETAIL = "To Public Project Detail",